requirements.txt
aiogram>=3.4
aiohttp>=3.9
requests>=2.31
fpdf2>=2.7
reportlab>=4.0
ruff>=0.5
black>=24.4
pytest>=8.2
pytest-asyncio>=0.23
sqlalchemy[asyncio]==2.0.30
aiosqlite==0.19.0
pydantic-settings==2.3.1
pyyaml==6.0.1
aiofiles==23.2.1
python-dotenv==1.0.1

